LPS-116098 Move tests to their corresponding directories
